pyproject.toml
[project]
name = "enterprise-knowledge-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.119.1",
    "ollama>=0.6.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
